--- a/cmd/tempo-cli/cmd-list-column.go
+++ b/cmd/tempo-cli/cmd-list-column.go
@@ -48,9 +48,13 @@ func (cmd *listColumnCmd) Run(ctx *globalOptions) error {
 		fmt.Printf("\n***************       rowgroup %d      ********************\n\n\n", i)
 
 		pages := cc.Pages()
-		numPages := cc.ColumnIndex().NumPages()
-		fmt.Println("Min Value of rowgroup", cc.ColumnIndex().MinValue(0).Bytes())
-		fmt.Println("Max Value of rowgroup", cc.ColumnIndex().MaxValue(numPages-1).Bytes())
+		idx, err := cc.ColumnIndex()
+		if err != nil {
+			return err
+		}
+		numPages := idx.NumPages()
+		fmt.Println("Min Value of rowgroup", idx.MinValue(0).Bytes())
+		fmt.Println("Max Value of rowgroup", idx.MaxValue(numPages-1).Bytes())
 
 		buffer := make([]parquet.Value, 10000)
 		for {