--- a/cmd/tempo-cli/cmd-view-pq-schema.go
+++ b/cmd/tempo-cli/cmd-view-pq-schema.go
@@ -51,8 +51,12 @@ func (cmd *viewSchemaCmd) Run(ctx *globalOptions) error {
 			path, _ := getNodePathByIndex(pf.Root(), "", cc.Column())
 
 			var size int64
-			for pg := 0; pg < cc.OffsetIndex().NumPages(); pg++ {
-				size += cc.OffsetIndex().CompressedPageSize(pg)
+			idx, err := cc.OffsetIndex()
+			if err != nil {
+				return err
+			}
+			for pg := 0; pg < idx.NumPages(); pg++ {
+				size += idx.CompressedPageSize(pg)
 			}
 
 			columnSizes[path] = columnSizes[path] + size