--- a/go.mod
+++ b/go.mod
@@ -110,7 +110,7 @@ require (
 	github.com/open-telemetry/opentelemetry-collector-contrib/pkg/ottl v0.89.0
 	github.com/open-telemetry/opentelemetry-collector-contrib/processor/filterprocessor v0.89.0
 	github.com/parquet-go/parquet-go v0.20.1-0.20231211122939-bec192ac8b06
-	github.com/stoewer/parquet-cli v0.0.5
+	github.com/stoewer/parquet-cli v0.0.7
 	go.opentelemetry.io/collector/config/configgrpc v0.89.0
 	go.opentelemetry.io/collector/config/confighttp v0.89.0
 	go.opentelemetry.io/collector/config/configtls v0.89.0