**What this PR does**:
Updates parquet-go to tip of main.

```
 (parquet-gogogo)$ benchstat before.txt after.txt 
name                                             old time/op    new time/op    delta
BackendBlockTraceQL/spanAttValNoMatch-8            8.95ms ± 1%    8.79ms ± 1%   -1.78%  (p=0.000 n=10+10)
BackendBlockTraceQL/spanAttIntrinsicNoMatch-8      8.44ms ± 1%    8.41ms ± 1%     ~     (p=0.063 n=9+9)
BackendBlockTraceQL/resourceAttValNoMatch-8        8.30ms ± 2%    8.31ms ± 1%     ~     (p=0.853 n=10+10)
BackendBlockTraceQL/resourceAttIntrinsicMatch-8    22.7ms ± 1%    22.3ms ± 1%   -1.81%  (p=0.000 n=10+9)
BackendBlockTraceQL/mixedValNoMatch-8               263ms ± 3%     254ms ± 3%   -3.46%  (p=0.000 n=10+10)
BackendBlockTraceQL/mixedValMixedMatchAnd-8        8.32ms ± 1%    8.30ms ± 1%     ~     (p=0.400 n=10+9)
BackendBlockTraceQL/mixedValMixedMatchOr-8          275ms ± 3%     269ms ± 3%   -2.40%  (p=0.005 n=10+10)
BackendBlockTraceQL/count-8                         488ms ±19%     491ms ±11%     ~     (p=0.796 n=10+10)
BackendBlockTraceQL/struct-8                        787ms ± 2%     785ms ± 6%     ~     (p=0.842 n=9+10)
BackendBlockTraceQL/||-8                            159ms ± 1%     155ms ± 1%   -2.12%  (p=0.000 n=10+10)
BackendBlockTraceQL/mixed-8                        29.6ms ± 2%    29.0ms ± 2%   -1.77%  (p=0.005 n=10+10)

name                                             old speed      new speed      delta
BackendBlockTraceQL/spanAttValNoMatch-8           101MB/s ± 1%   103MB/s ± 1%   +1.81%  (p=0.000 n=10+10)
BackendBlockTraceQL/spanAttIntrinsicNoMatch-8     136MB/s ± 1%   137MB/s ± 1%     ~     (p=0.059 n=9+9)
BackendBlockTraceQL/resourceAttValNoMatch-8      13.2MB/s ± 2%  13.2MB/s ± 1%     ~     (p=0.864 n=10+10)
BackendBlockTraceQL/resourceAttIntrinsicMatch-8   618MB/s ± 1%   629MB/s ± 1%   +1.84%  (p=0.000 n=10+9)
BackendBlockTraceQL/mixedValNoMatch-8            4.36MB/s ± 3%  4.52MB/s ± 3%   +3.58%  (p=0.000 n=10+10)
BackendBlockTraceQL/mixedValMixedMatchAnd-8      14.1MB/s ± 1%  14.1MB/s ± 1%     ~     (p=0.461 n=10+9)
BackendBlockTraceQL/mixedValMixedMatchOr-8       55.2MB/s ± 3%  56.5MB/s ± 3%   +2.47%  (p=0.005 n=10+10)
BackendBlockTraceQL/count-8                      31.2MB/s ±16%  31.0MB/s ±10%     ~     (p=0.796 n=10+10)
BackendBlockTraceQL/struct-8                     4.65MB/s ± 2%  4.66MB/s ± 5%     ~     (p=0.825 n=9+10)
BackendBlockTraceQL/||-8                         97.3MB/s ± 1%  99.4MB/s ± 1%   +2.16%  (p=0.000 n=10+10)
BackendBlockTraceQL/mixed-8                       470MB/s ± 2%   479MB/s ± 2%   +1.80%  (p=0.005 n=10+10)

name                                             old MB_io/op   new MB_io/op   delta
BackendBlockTraceQL/spanAttValNoMatch-8              0.90 ± 0%      0.90 ± 0%     ~     (all equal)
BackendBlockTraceQL/spanAttIntrinsicNoMatch-8        1.15 ± 0%      1.15 ± 0%     ~     (all equal)
BackendBlockTraceQL/resourceAttValNoMatch-8          0.11 ± 0%      0.11 ± 0%     ~     (all equal)
BackendBlockTraceQL/resourceAttIntrinsicMatch-8      14.0 ± 0%      14.0 ± 0%     ~     (all equal)
BackendBlockTraceQL/mixedValNoMatch-8                1.15 ± 0%      1.15 ± 0%     ~     (all equal)
BackendBlockTraceQL/mixedValMixedMatchAnd-8          0.12 ± 0%      0.12 ± 0%     ~     (all equal)
BackendBlockTraceQL/mixedValMixedMatchOr-8           15.2 ± 0%      15.2 ± 0%     ~     (all equal)
BackendBlockTraceQL/count-8                          15.1 ± 0%      15.1 ± 0%     ~     (all equal)
BackendBlockTraceQL/struct-8                         3.66 ± 0%      3.66 ± 0%   +0.03%  (p=0.000 n=10+10)
BackendBlockTraceQL/||-8                             15.4 ± 0%      15.4 ± 0%     ~     (all equal)
BackendBlockTraceQL/mixed-8                          13.9 ± 0%      13.9 ± 0%     ~     (all equal)

name                                             old alloc/op   new alloc/op   delta
BackendBlockTraceQL/spanAttValNoMatch-8            5.09MB ± 2%    4.60MB ± 2%   -9.46%  (p=0.000 n=10+10)
BackendBlockTraceQL/spanAttIntrinsicNoMatch-8      4.50MB ± 2%    4.45MB ± 1%   -1.10%  (p=0.011 n=10+10)
BackendBlockTraceQL/resourceAttValNoMatch-8        4.41MB ± 2%    4.42MB ± 1%     ~     (p=0.796 n=10+10)
BackendBlockTraceQL/resourceAttIntrinsicMatch-8    5.26MB ± 6%    5.28MB ± 6%     ~     (p=0.489 n=9+9)
BackendBlockTraceQL/mixedValNoMatch-8              5.63MB ±31%    5.40MB ±28%     ~     (p=0.481 n=10+10)
BackendBlockTraceQL/mixedValMixedMatchAnd-8        4.46MB ± 1%    4.42MB ± 2%     ~     (p=0.052 n=10+10)
BackendBlockTraceQL/mixedValMixedMatchOr-8         5.91MB ±30%    4.45MB ± 0%  -24.67%  (p=0.000 n=10+8)
BackendBlockTraceQL/count-8                         279MB ± 7%     280MB ± 6%     ~     (p=0.739 n=10+10)
BackendBlockTraceQL/struct-8                       58.4MB ± 5%    57.4MB ± 7%     ~     (p=0.133 n=9+10)
BackendBlockTraceQL/||-8                           83.6MB ± 2%    84.0MB ± 2%     ~     (p=0.353 n=10+10)
BackendBlockTraceQL/mixed-8                        4.95MB ±11%    5.01MB ±11%     ~     (p=0.739 n=10+10)

name                                             old allocs/op  new allocs/op  delta
BackendBlockTraceQL/spanAttValNoMatch-8             64.1k ± 0%     64.0k ± 0%   -0.11%  (p=0.000 n=10+10)
BackendBlockTraceQL/spanAttIntrinsicNoMatch-8       64.0k ± 0%     63.9k ± 0%   -0.06%  (p=0.000 n=10+10)
BackendBlockTraceQL/resourceAttValNoMatch-8         64.0k ± 0%     63.9k ± 0%   -0.03%  (p=0.000 n=10+10)
BackendBlockTraceQL/resourceAttIntrinsicMatch-8     67.8k ± 0%     67.8k ± 0%   -0.05%  (p=0.000 n=9+9)
BackendBlockTraceQL/mixedValNoMatch-8               65.7k ± 0%     65.6k ± 0%   -0.11%  (p=0.000 n=10+10)
BackendBlockTraceQL/mixedValMixedMatchAnd-8         64.1k ± 0%     64.0k ± 0%   -0.03%  (p=0.000 n=8+10)
BackendBlockTraceQL/mixedValMixedMatchOr-8          65.7k ± 0%     65.7k ± 0%   -0.10%  (p=0.000 n=10+8)
BackendBlockTraceQL/count-8                         1.30M ± 0%     1.30M ± 0%     ~     (p=0.366 n=9+10)
BackendBlockTraceQL/struct-8                         211k ± 1%      211k ± 1%     ~     (p=0.243 n=9+10)
BackendBlockTraceQL/||-8                             451k ± 0%      451k ± 0%   -0.01%  (p=0.000 n=10+8)
BackendBlockTraceQL/mixed-8                         65.9k ± 0%     65.9k ± 0%   +0.06%  (p=0.000 n=10+10)
```